refactor(validate): delete dead markCaptured/Captured analysis
markCapturedBindings wrote ValidatedLetBinding.Captured, which had zero
production readers — the B1 consumer from PR #604 never landed. Resolves
the TODO.md "delete or unify" item via option (a) (delete):

  - remove validate_capture.go (markCapturedBindings)
  - remove the ValidatedLetBinding.Captured field
  - remove its 5 call sites in validate_let.go
  - remove validate_capture_test.go (TestMarkCapturedBindings_*)

The live siblings are untouched: markEscaped/.Escapes (read at
compile_let.go:240 for let-lambda inlining) and the unrelated
bodyReferencesCaptureOperator (call/cc detection, read by
frame_reclaim_build.go).

The deleted test file also held the package's shared test helpers
(call/symRef/lam/makeTestEnvAndBindings/setBang, used across sibling
test files); relocated them to sharedtest_test.go so they no longer
live inside a feature-specific file.
