+++ TODO.md
@@ -169,7 +169,7 @@ Directions documents — identify prioritized capability extensions. Priority se
 - [ ] **Environment frame slimming** [Performance]: Reduce `EnvironmentFrame` struct for closure bodies that only need local bindings. `plans/PERFORMANCE.md`
 - [ ] **B3 effective capture refinement** [Performance, Research]: Propagate B2 escape results back into B1 capture status. A binding marked `Captured` by B1 is effectively non-captured if every lambda that references it is stored in a non-escaping binding (B2). Cross-binding analysis over B1+B2 results.
 - [ ] **`PrimitiveSpec` capture-safety capability field** [Performance, M, follow-up to escape-frame-validation]: The escape-frame classifier's Layer C (`internal/validate/frame_reclaim_build.go`) decides whether a *primitive* callee is capture-safe via a hand-maintained, sound-by-default name whitelist (`captureSafePrimitiveNames` — `+`, `cons`, `<`, … contribute no edge; unlisted ⇒ unsafe). A primitive is capture-*unsafe* iff it can invoke a Scheme procedure that captures (`apply`, `map`, `for-each`, `call-with-values`, `dynamic-wind`, `with-exception-handler`, `sort`, …). The principled replacement is a `PrimitiveSpec.InvokesProcedure` (or `CapturesContinuation`) capability field so each primitive — including extension primitives — self-declares, instead of a central name list that silently under-covers extensions. **Gate:** only worth building once the sibling escape-gated plan's Phase 2 measurement shows the classifier's precision matters (the whitelist is sound regardless; this is a coverage/scalability dial). Weight false-positives (declaring a capturing primitive safe) as unacceptable per `feedback_annotation_stability.md`. Q-1 of `plans/2026-06-12-escape-frame-validation-impl.md`.
-- [ ] **`markCaptured`/`.Captured` is dead code — delete or unify** [Tech debt, S, follow-up to escape-frame-validation]: `markCapturedBindings` (`internal/validate/validate_capture.go`) writes `ValidatedLetBinding.Captured`, which has **zero production readers** (only `validate_capture_test.go` asserts it) — the B1 consumer from PR #604 never landed (it's the unbuilt half of the B3 item above). `markEscaped`/`.Escapes`, by contrast, is live (read at `compile_let.go:240` for let-lambda inlining). Surfaced while building the escape-gated `FrameReclaimable`/`MayCapture` classifier (`plans/2026-06-12-escape-frame-validation-impl.md`), which is *additive* and deliberately does not touch these. **Decision needed:** (a) delete `markCaptured`+`.Captured`+its tests as dead since PR #604, or (b) keep, and unify all three capture/escape predicate-layers (markCaptured, markEscaped, the new fail-safe `MayCapture`) onto the single `WalkBindingRefs` traversal they already share — gated on whether B1/B3 ever gets a consumer. Folding the fail-safe predicate in with the two best-effort ones risks regressing the inlining contract (`compile_let.go:240`), so unify only with a substitution check across all three call sites.
+- [x] **`markCaptured`/`.Captured` is dead code — delete or unify** [Tech debt, S, follow-up to escape-frame-validation, Done]: Resolved via option (a) — deleted `markCapturedBindings` (`internal/validate/validate_capture.go`), the `ValidatedLetBinding.Captured` field, its 5 call sites in `validate_let.go`, and `validate_capture_test.go`. Shared test helpers from that file (`call`/`symRef`/`lam`/`makeTestEnvAndBindings`/… still used by sibling test files) were relocated to `internal/validate/sharedtest_test.go`. The live `markEscaped`/`.Escapes` path (read at `compile_let.go:240` for let-lambda inlining) and the unrelated live `bodyReferencesCaptureOperator` (call/cc detection, read by `frame_reclaim_build.go`) were left untouched. Unification (option b) was rejected: with B1/B3 never built there is no second consumer to unify against, and folding a fail-safe predicate in with the best-effort one risked regressing the inlining contract.
 - [ ] **Benchmark coverage gaps** [Performance, S-M]: No benchmarks for compiler, expander (syntax-rules expansion), library import resolution, or continuation capture/restore cycle. Existing benchmarks cover VM dispatch, fibonacci, tokenizer, parser, environment, and symbol interning.
 - [ ] **Fused lexing/parsing** [Performance, Research]: Flap paper (PLDI 2023) — fuse tokenizer and parser into single character-level pass, eliminating per-token heap allocation. Gated on profiling confirming tokenizer is a bottleneck. `plans/PERFORMANCE.md`
 - [ ] **Inline-budget guard for `checkStackSize` and similar hot-path wrappers** [Performance, S]: `checkStackSize` (`machine/machine_context.go:1185`) is split from `reportStackOverflow` specifically to stay under Go's 80-cost inline budget (currently 67). A future innocuous edit could push it over and silently regress the VM hot path (the Gabriel suite would catch it, but only post-hoc and noisily). Write a test that runs `go build -gcflags='-m=2' ./machine/` and asserts `"can inline (*MachineContext).checkStackSize"` appears in the output. ~30 LOC test infrastructure; reusable for future hot-path wrappers. Surfaced by Finding 5 / PR #734 type-design review.
