+++ internal/validate/validated_forms.go
@@ -303,14 +303,12 @@ func (p LetKind) Sequential() bool {
 
 // ValidatedLetBinding represents a single (name init-expr) binding pair.
 // Mutable is true if the binding is targeted by set! in the body.
-// Captured is true if the binding is referenced from inside an escaping closure.
 // Escapes is true if the binding is referenced in a non-call position.
 type ValidatedLetBinding struct {
-	Name     *syntax.SyntaxSymbol
-	Init     ValidatedExpr
-	Mutable  bool
-	Captured bool
-	Escapes  bool
+	Name    *syntax.SyntaxSymbol
+	Init    ValidatedExpr
+	Mutable bool
+	Escapes bool
 }
 
 // ValidatedLet represents all four R7RS binding forms: let, let*, letrec,
